backend/requirements.txt
# Django.
Django==3.1.1
djangorestframework==3.11.1

# Code quality.
black==20.8b1
coverage==5.3

# Static file serving.
WhiteNoise==5.2.0
